requirements.txt
requests==2.32.3
python-dotenv==0.9.1